ledger: fix the index calculation when translating descriptor to policy
e571f4fafc54384d22142e325897f8a90393cf7e (that didn't came with any unit
test demonstrating the fixed behaviour, which would have caught the
present bug) fixed how we could pass more than once the same pubkey to
the Ledger if an xpub is reused across paths in the descriptor. However
it still was using the index as if we were using duplicating pubkeys
thereby resulting in pubkey indexes in the wallet policy that were out
of bounds.