Add utils module with bip32_derivation filter
Some devices like Bitbox, sign only the first bip32 derivation
that matches the master fingerprint of the device even if
a key signed already for this particular derivation or if
inputs have differents bip32 derivation order, causing problem
if multiple spending paths bip32 derivations are present.

With the struct Bip32DerivationFilter, it is possible to ignore
specific derivations the time to sign the psbt.